Refactors unittest directory
