+++ Makefile
@@ -1,0 +1,4 @@
+install:
+	mkdir database
+	mkdir logs
+	pip install -e . && python setup.py develop
