+++ test/http_requests/delete_comment_local.http
